rename API to XKCD
