+++ internal-xkcd/adapters/httpserver/handlers/scrape_test.go
@@ -8,10 +8,10 @@ import (
 	"fmt"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"myapp/internal-api/config"
-	"myapp/internal-api/core/domain"
-	"myapp/internal-api/core/port/mock"
-	mock2 "myapp/internal-api/core/util/mock"
+	"myapp/internal-xkcd/config"
+	"myapp/internal-xkcd/core/domain"
+	"myapp/internal-xkcd/core/port/mock"
+	mock2 "myapp/internal-xkcd/core/util/mock"
 	"net/http"
 	"net/http/httptest"
 	"strings"
