+++ internal-xkcd/core/port/mock/scraper.go
