+++ internal-xkcd/core/service/scrape.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"myapp/internal-api/config"
-	"myapp/internal-api/core/domain"
-	"myapp/internal-api/core/port"
-	"myapp/internal-api/core/util"
+	"myapp/internal-xkcd/config"
+	"myapp/internal-xkcd/core/domain"
+	"myapp/internal-xkcd/core/port"
+	"myapp/internal-xkcd/core/util"
 	"net/http"
 	"sync"
 )
