+++ internal-xkcd/core/service/scrape_test.go
@@ -3,11 +3,11 @@ package service
 import (
 	"context"
 	"github.com/stretchr/testify/assert"
-	"myapp/internal-api/config"
-	"myapp/internal-api/core/domain"
-	"myapp/internal-api/core/port/mock"
-	"myapp/internal-api/core/util"
-	mockUtil "myapp/internal-api/core/util/mock"
+	"myapp/internal-xkcd/config"
+	"myapp/internal-xkcd/core/domain"
+	"myapp/internal-xkcd/core/port/mock"
+	"myapp/internal-xkcd/core/util"
+	mockUtil "myapp/internal-xkcd/core/util/mock"
 	"slices"
 	"sort"
 	"sync"
